Moved all styles to a separate file styles.js
